+++ utils/toastConfig.tsx
@@ -1,0 +1,17 @@
+import React from "react";
+import { BaseToast } from "react-native-toast-message";
+
+const toastConfig = {
+  success: (props: any) => (
+    <BaseToast
+      {...props}
+      style={{ borderLeftColor: "pink" }}
+      contentContainerStyle={{ paddingHorizontal: 24 }}
+      text1Style={{
+        fontSize: 15,
+        fontWeight: "400",
+      }}
+    />
+  ),
+};
+export default toastConfig;
